Put the Alert into a component. Changed App.vue to use the component
