+++ src/App.vue
@@ -6,15 +6,7 @@
     <NavbarComponent />
 
     <!-- ALERT -->
-    <div class="text-center py-4 lg:px-4" id="upcoming-alert">
-      <div class="p-2 bg-accent-2 items-center leading-none lg:rounded-full flex lg:inline-flex" role="alert">
-        <span class="flex rounded-full bg-primary text-white uppercase px-2 py-1 text-xs font-bold font-display mr-3">New</span>
-        <span class="font-body mr-2 text-left flex-auto" style="font-color: #12418f;">Join our <a class="underline" href="https://discord.gg/wVwv5Pw" target="_blank">Discord</a> to learn about meetings. Need to download Kali? Click <a class="underline" href="https://kali.org/downloads/" target="_blank">here</a>!</span>        <svg class="fill-current opacity-75 h-4 w-4" role="button" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" v-on:click="hideAlert()">
-          <title>Close</title>
-          <path d="M14.348 14.849a1.2 1.2 0 0 1-1.697 0L10 11.819l-2.651 3.029a1.2 1.2 0 1 1-1.697-1.697l2.758-3.15-2.759-3.152a1.2 1.2 0 1 1 1.697-1.697L10 8.183l2.651-3.031a1.2 1.2 0 1 1 1.697 1.697l-2.758 3.152 2.758 3.15a1.2 1.2 0 0 1 0 1.698z"/>
-        </svg>
-      </div>
-    </div>
+    <Alert type="New" message="Join our <a class='underline' href='https://discord.gg/wVwv5Pw' target='_blank'>Discord</a> to learn about meetings. Need to download Kali? Click <a class='underline' href='https://kali.org/downloads/' target='_blank'>here</a>!" />
 
     <!-- PAGE RENDER -->
     <router-view />
@@ -49,22 +41,17 @@
 <script>
 import NavbarComponent from './components/NavbarComponent.vue'
 import FooterComponent from './components/FooterComponent.vue'
+import Alert from './components/AlertComponent.vue'
 
 export default {
   name: 'app',
   components: {
     NavbarComponent,
-    FooterComponent
+    FooterComponent,
+    Alert
   },
   data () {
     return {
-      visible: true
-    }
-  },
-  methods: {
-    hideAlert: function () {
-      let alert = document.getElementById('upcoming-alert')
-      if (alert.style.display !== 'none') alert.style.display = 'none'
     }
   }
 }
